@ throws tag should also be aligned in JSDoc blocks
When using @throws, without this rule eslint complains if your statement is aligned. --fix would also revert the alignment. Now it will actually let you do it correctly and align the JSDoc.

Example of what happens currently:

```
/**
 * @param {string}                    interpolatedString The interpolation string to be parsed.
 * @param {Record<string, WPElement>} conversionMap      The map used to convert the string to
 *                                                       a react element.
 * @throws {TypeError}
 * @return {WPElement}  A wp element.
 */
```

This example looks fine but throws and return arent actually aligned with the param tags, which looks strange especially because the param tags are aligned.

With the new rule it would look like this instead:
```
 *
 * @param {string}                    interpolatedString The interpolation string to be parsed.
 * @param {Record<string, WPElement>} conversionMap      The map used to convert the string to
 *                                                       a react element.
 * @throws {TypeError}
 * @return {WPElement}  A wp element.
 */
```